last version had a clear scores function that deleted the entire array on firestore. Fixed it in order to reset the values to 0 instead of deleting the entire array.
also now each screen will get the current user (just need to make sure the getCurrentUser function gets called only if not null. that's why I am only using it on screens that we can access after logging in only for now. if i'm gonna use the function on the rest of the screen i'll need a try and catch block).